requirements.txt
Django==2.0.5
djangorestframework==3.8.2
mysqlclient==1.3.12
PyJWT==1.6.1
pytz==2018.4
